requirements.txt
aiofiles==0.6.0
aniso8601==7.0.0
apispec==4.0.0
apistar==0.7.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docopt==0.6.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==2.0.0
greenlet==1.0a1
h11==0.9.0
httptools==0.0.13
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
promise==2.3
psycopg2-binary==2.8.6
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
responder==2.0.5
rfc3986==1.4.0
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.12.13
typesystem==0.2.4
urllib3==1.26.2
uvicorn==0.10.9
uvloop==0.14.0
websockets==8.1
whitenoise==5.2.0
